pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pyservice"
authors = [
    {name = "Max Dubrovin"},
]

description = "Common tools for creating microservice."
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pylint",
    "pytz",
    "aiofiles",
    "celery",
    "pydantic",
    "pydantic_extra_types",
    "pydantic_settings",
    "PyYAML",
    "pika",
    "toml",
    "jinja2==3.1.4",
    "seqlog==0.3.31",
    "prettytable==3.12.0",

    "djangorestframework==3.15.2",
    "python-keycloak==5.3.1",



    "tgs-client @ git+https://pack:F6VxZN5xuHrd5fsdK9_V@git.cebb.pro/ias-phoenix-native/submodules/telegram-service-client.git@_package",
]
